fixed a few compiler warnings
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4698 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -1825,7 +1825,7 @@ static size_t sphReadThrottled ( int iFD, void * pBuf, size_t iCount, ThrottleSt
 	}
 
 	sphThrottleSleep ( pThrottle );
-	return sphRead ( iFD, pBuf, iCount );
+	return (size_t)sphRead ( iFD, pBuf, iCount ); // FIXME? we sure this is under 2gb?
 }
 
 void SafeClose ( int & iFD )
